+++ src/EventSubscriber/ReaderBeforeRenderEventSubscriber.php
@@ -13,9 +13,10 @@ use HeimrichHannot\ReaderBundle\Model\ReaderConfigElementModel;
 use HeimrichHannot\ReaderBundle\Registry\ReaderConfigElementRegistry;
 use HeimrichHannot\SyndicationTypeBundle\ConfigElementType\SyndicationConfigElementType;
 use HeimrichHannot\SyndicationTypeBundle\SyndicationType\ExportSyndicationHandler;
+use Psr\Container\ContainerInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
-class ReaderBeforeRenderEventSubscriber implements EventSubscriberInterface
+class ReaderBeforeRenderEventSubscriber extends AbstractServiceSubscriber implements EventSubscriberInterface
 {
     /**
      * @var ExportSyndicationHandler
@@ -25,14 +26,18 @@ class ReaderBeforeRenderEventSubscriber implements EventSubscriberInterface
      * @var ReaderConfigElementRegistry
      */
     protected $configElementRegistry;
+    /**
+     * @var ContainerInterface
+     */
+    protected $container;
 
     /**
      * ReaderBeforeRenderEventSubscriber constructor.
      */
-    public function __construct(ExportSyndicationHandler $exportSyndicationHandler, ReaderConfigElementRegistry $configElementRegistry)
+    public function __construct(ContainerInterface $container, ExportSyndicationHandler $exportSyndicationHandler)
     {
         $this->exportSyndicationHandler = $exportSyndicationHandler;
-        $this->configElementRegistry = $configElementRegistry;
+        $this->container = $container;
     }
 
     public static function getSubscribedEvents()
@@ -44,6 +49,10 @@ class ReaderBeforeRenderEventSubscriber implements EventSubscriberInterface
 
     public function onReaderBeforeRender(ReaderBeforeRenderEvent $event): void
     {
+        if (!$this->container->has('HeimrichHannot\ReaderBundle\Registry\ReaderConfigElementRegistry')) {
+            return;
+        }
+
         $readerConfigElements = ReaderConfigElementModel::findBy([
             'tl_reader_config_element.pid=?',
             'tl_reader_config_element.type=?',
@@ -57,7 +66,7 @@ class ReaderBeforeRenderEventSubscriber implements EventSubscriberInterface
         }
 
         /** @var SyndicationConfigElementType $configElementType */
-        $configElementType = $this->configElementRegistry->getReaderConfigElementType(SyndicationConfigElementType::getType());
+        $configElementType = $this->container->get(ReaderConfigElementRegistry::class)->getReaderConfigElementType(SyndicationConfigElementType::getType());
 
         if (!$configElementType) {
             return;
@@ -70,5 +79,12 @@ class ReaderBeforeRenderEventSubscriber implements EventSubscriberInterface
             );
             $this->exportSyndicationHandler->exportByContext($context);
         }
+    }
+
+    public static function getSubscribedServices()
+    {
+        return [
+            '?HeimrichHannot\ReaderBundle\Registry\ReaderConfigElementRegistry',
+        ];
     }
 }